Split up files and create observational handlers folder
