+++ causal_pyro/observational/internals.py
@@ -1,6 +1,7 @@
 from typing import Optional, TypeVar
 
 import pyro
+import pyro.distributions
 import torch
 
 from causal_pyro.observational.ops import AtomicObservation, observe
@@ -14,7 +15,7 @@ T = TypeVar("T")
 @observe.register(torch.Tensor)
 def _observe_deterministic(
     rv: T, obs: Optional[AtomicObservation[T]] = None, **kwargs
-) -> T:
+):
     """
     Observe a tensor in a probabilistic program.
     """
@@ -37,19 +38,7 @@ def _observe_distribution(
         raise ValueError("name must be specified when observing a distribution")
 
     if callable(obs):
-        # rv_ = pyro.sample(name, rv, obs=None, **kwargs)
-        # obs = obs(rv_)
-        # return observe(rv_, obs, name=name, **kwargs)
-        raise NotImplementedError(
-            "Observing a distribution with a callable is not yet supported"
-        )
-
-    if isinstance(obs, tuple):
-        # for i, o in enumerate(obs):
-        #     observe(rv, o, name=f"{name}[{i}]", **kwargs)
-        raise NotImplementedError(
-            "Observing a distribution with a tuple is not yet supported"
-        )
+        raise NotImplementedError("Dependent observations are not yet supported")
 
     return pyro.sample(name, rv, obs=obs, **kwargs)
